moved load_data function to utilities
